requirements.txt
asgiref==3.6.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
DateTime==5.1
Django==3.2.15
django-cors-headers==3.14.0
djangorestframework==3.13.1
idna==3.4
isodate==0.6.1
minio==7.1.14
Pillow==9.5.0
psutil==5.9.5
psycopg2==2.9.6
pybase64==1.2.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.27.0
requires.io==0.2.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==4.1.2
zope.interface==6.0